+++ docs/source/changelog.rst
@@ -5,8 +5,13 @@ Releases
 Version 1.0 Beta 2 (2013-06-20)
 ===============================
 
+- Renamed function ``wsgify_django`` to
+  :func:`django_pastedeploy_settings.get_configured_django_wsgi_app`.
 - Made the :doc:`Nose plugin <testing>` work with multiple databases as
   supported by Django >= 1.3.
+- Renamed command-line argument for the PasteDeploy configuration file in the
+  :doc:`Nose plugin <testing>` from ``--with-django-wsgified`` to
+  ``--paste-deploy-config``.
 
 
 Version 1.0 Beta 1 (2013-06-19)